Split errors up and moved caret completion
