+++ src/main/scala/parsley/internal/errors/ParseError.scala
@@ -1,17 +1,15 @@
 package parsley.internal.errors
 
 import ParseError.Unknown
-import Raw.Unprintable
-import scala.util.matching.Regex
 
-private [internal] sealed abstract class ParseError {
+private [internal] sealed trait ParseError {
     val offset: Int
     val col: Int
     val line: Int
 
     def withHints(hints: Set[ErrorItem]): ParseError
     def giveReason(reason: String): ParseError
-    def pretty(sourceName: Option[String])(implicit helper: LineBuilder): String
+    private [internal] def pretty(sourceName: Option[String])(implicit helper: LineBuilder): String
 
     protected final def posStr(sourceName: Option[String]): String = {
         val scopeName = sourceName.fold("")(name => s"In file '$name' ")
@@ -31,7 +29,7 @@ private [internal] sealed abstract class ParseError {
         val topStr = posStr(sourceName)
         val (line, caret) = helper.getLineWithCaret(offset)
         val info = infoLines.filter(_.nonEmpty).mkString("\n  ")
-        s"$topStr:\n  ${if (info.isEmpty) Unknown else info}\n  >${line}\n  >${caret}"
+        s"$topStr:\n  ${if (info.isEmpty) Unknown else info}\n  >${line}\n  >${" " * caret}^"
     }
 }
 // The reasons here are lightweight, two errors can merge their messages, but messages do not get converted to hints
@@ -60,36 +58,4 @@ private [internal] object ParseError {
     val Unknown = "unknown parse error"
     val NoReason = Set.empty[String]
     val NoItems = Set.empty[ErrorItem]
-}
-
-private [internal] sealed trait ErrorItem {
-    val msg: String
-}
-private [internal] object ErrorItem {
-    def higherPriority(e1: ErrorItem, e2: ErrorItem): ErrorItem = (e1, e2) match {
-        case (EndOfInput, _) => EndOfInput
-        case (_, EndOfInput) => EndOfInput
-        case (e: Desc, _) => e
-        case (_, e: Desc) => e
-        case (Raw(r1), Raw(r2)) => if (r1.length >= r2.length) e1 else e2
-    }
-}
-private [internal] case class Raw(cs: String) extends ErrorItem {
-    // This could be marked threadUnsafe in Scala 3?
-    override lazy val msg = cs match {
-        case "\n"            => "newline"
-        case "\t"            => "tab"
-        case " "             => "space"
-        case Unprintable(up) => f"unprintable character (\\u${up.head.toInt}%04X)"
-        // Do we want this only in unexpecteds?
-        case cs              => "\"" + cs.takeWhile(c => c != '\n' && c != ' ') + "\""
-    }
-}
-private [internal] object Raw {
-    val Unprintable: Regex = "(\\p{C})".r
-    def apply(c: Char): Raw = new Raw(s"$c")
-}
-private [internal] case class Desc(msg: String) extends ErrorItem
-private [internal] case object EndOfInput extends ErrorItem {
-    override val msg = "end of input"
 }
